Moved functions of interest into their own file to avoid name collisions during testing.
